@@ -10,16 +10,16 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~13.0.2",
+    "@angular/animations": "^14.2.6",
     "@angular/cdk": "^13.0.2",
-    "@angular/common": "~13.0.2",
-    "@angular/compiler": "~13.0.2",
-    "@angular/core": "~13.0.2",
+    "@angular/common": "^14.2.6",
+    "@angular/compiler": "^14.2.6",
+    "@angular/core": "^14.2.6",
     "@angular/flex-layout": "^13.0.0-beta.38",
-    "@angular/forms": "~13.0.2",
-    "@angular/platform-browser": "~13.0.2",
-    "@angular/platform-browser-dynamic": "~13.0.2",
-    "@angular/router": "~13.0.2",
+    "@angular/forms": "^14.2.6",
+    "@angular/platform-browser": "^14.2.6",
+    "@angular/platform-browser-dynamic": "^14.2.6",
+    "@angular/router": "^14.2.6",
     "@ngneat/transloco": "^3.1.1",
     "@ngneat/transloco-locale": "^3.0.1",
     "@stomp/rx-stomp": "^1.1.4",
@@ -32,9 +32,9 @@
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~13.0.3",
-    "@angular/cli": "^13.0.3",
-    "@angular/compiler-cli": "~13.0.2",
+    "@angular-devkit/build-angular": "^14.2.6",
+    "@angular/cli": "^14.2.6",
+    "@angular/compiler-cli": "^14.2.6",
     "@types/jasmine": "~3.10.2",
     "@types/node": "^16.11.10",
     "jasmine-core": "~3.10.1",
@@ -43,6 +43,6 @@
     "karma-coverage": "~2.0.3",
     "karma-jasmine": "~4.0.1",
     "karma-jasmine-html-reporter": "^1.7.0",
-    "typescript": "~4.4.0"
+    "typescript": "~4.8.4"
   }
 }
